\name{BayesPortfolio-package}
\alias{BayesPortfolio-package}
\alias{BayesPortfolio}
\docType{package}
\title{
\packageTitle{BayesPortfolio}
Factor Investing: A Bayesian Hierarchical Approach.
}
\description{
\packageDescription{BayesPortfolio}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesPortfolio}
\packageIndices{BayesPortfolio}
This package implements the method from Factor Investing: A Bayesian Hierarchical Approach. It provide a seemingly regression model with hierarchical prior to model multiple assets, predict stock returns and estimate covariance matrix. Thus build mean-variance efficient portfolio.
}
\author{
Jingyu He, Gavin Guanhao Feng
}
\references{
Feng, Guanhao and He, Jingyu, Factor Investing: A Bayesian Hierarchical Approach (July 19, 2020). Available at SSRN: https://ssrn.com/abstract=3326617 or http://dx.doi.org/10.2139/ssrn.3326617
}
\keyword{ package }
\seealso{
}
\examples{
}
